FIX: avoid using magic port within UT;
