+++ pkg/tests/utils/network_test.go
@@ -1,0 +1,47 @@
+// Copyright 2021 - 2022 Matrix Origin
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package utils
+
+import (
+	"strconv"
+	"testing"
+
+	"github.com/stretchr/testify/require"
+)
+
+func TestGetAvailablePort(t *testing.T) {
+	defer func() {
+		curPort = 10000
+	}()
+
+	curPort = maxPort
+	_, err := GetAvailablePort("127.0.0.1")
+	require.Error(t, err)
+
+	curPort = maxPort - 1
+	port, err := GetAvailablePort("127.0.0.1")
+	require.NoError(t, err)
+	require.Equal(t, port, strconv.Itoa(maxPort))
+}
+
+func TestGetAddressBatch(t *testing.T) {
+	defer func() {
+		curPort = 10000
+	}()
+
+	addrs, err := GetAddressBatch("127.0.0.1", 1)
+	require.NoError(t, err)
+	require.Equal(t, 1, len(addrs))
+}
